Added pipeline execution endpoint to controller. Moved start to server.js.
