+++ config/nodes/statLinks.node.json
@@ -1,5 +1,4 @@
 {
-  "id": 777,
   "name": "test.statsLink",
   "uuid": "cacc1bea-d0d8-4c0d-8121-10c8f9f7474a",
   "url": "http://localhost:3001/num_stats",
