+++ config/pipelines/numberStats.ppln.json
@@ -6,7 +6,6 @@
       "name": "Stat links",
       "description": "Get the list of stat urls for a list of numbers",
       "nodeUUID": "cacc1bea-d0d8-4c0d-8121-10c8f9f7474a",
-      "params": {},
       "data": {},
       "extract": {
         "links": "array:"
@@ -16,7 +15,6 @@
       "name": "Stat Step",
       "description": "Calculate a statistic of a list of numbers",
       "nodeUUID":  "b5c621d9-7540-47d3-87d5-2d5fbe46380d",
-      "params": {},
       "data": {
         "stats": ["sum", "avg", "min", "max"]
       },
